+++ FriendlyTOM/FriendlyTOM/UserControls/BookingsUserControl.xaml.cs
@@ -113,12 +113,17 @@ namespace FriendlyTOM.UserControls
         {
             if (selectedBooking != null)
             {
-                foreach (IBooking booking in bookingsDataGrid.SelectedItems)
+                MessageBoxResult doDelete = MessageBox.Show("Delete selected booking(s)?", "Confirm deletion",
+                    MessageBoxButton.YesNo, MessageBoxImage.Warning);
+                if (doDelete == MessageBoxResult.Yes)
                 {
-                    bookingController.DeleteBooking(booking);
+                    foreach (IBooking booking in bookingsDataGrid.SelectedItems)
+                    {
+                        bookingController.DeleteBooking(booking);
+                    }
+                    bookingsDataGrid.SelectedItem = null;
+                    refreshDataGrid();
                 }
-                bookingsDataGrid.SelectedItem = null;
-                refreshDataGrid();
             }
         }
 
@@ -339,10 +344,18 @@ namespace FriendlyTOM.UserControls
             }
         }
 
-        private void calculatePayments_Click(object sender, RoutedEventArgs e)
+        private void createPayments_Click(object sender, RoutedEventArgs e)
         {
-            bookingController.CalculatePaymentsForBooking(selectedBooking);
-            refreshDataGrid();
+            if (selectedBooking == null)
+            {
+                MessageBox.Show("Please select a booking first!");
+            }
+            else
+            {
+                updateExistingBooking();
+                bookingController.CreatePaymentsForBooking(selectedBooking);
+                refreshDataGrid();
+            }
         }
         #endregion
     }
